Refactored code structure a bit for easier maintenance.
Moved the memory map-related code into its own module, which is
itself a submodule of a newly created process module. The idea
could be to hide the memory details from the user and only
expose it at a higher level as process information. Processes
could be queries for memory information via function calls.

Maybe the proc_utils thing should be moved into a separate crate.